+++ src/main/java/kr/co/pawong/pwbe/lostPost/adapter/in/api/dto/request/LostPostUpdate.java
@@ -1,0 +1,5 @@
+package kr.co.pawong.pwbe.lostPost.adapter.in.api.dto.request;
+
+public class LostPostUpdate {
+
+}
