--- src/main/java/kr/co/pawong/pwbe/lostPost/domain/LostPostUpdate.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostPost.domain;
-
-public class LostPostUpdate {
-
-}
